backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.7.0
channels==4.2.0
Pillow==11.1.0
redis==5.2.1